simplify alive_on and existed_on logic
Changed this in the process of fixing a bug where pairing.eggs()
wouldn't play well with existed_on() or alive_on() because they
were both doing filtering joins on the event table. The simpler
solution is to use with_dates() to get the birth/death/etc dates
and then filter based on whether these are null or not.