first version of pairing-based breeding report done
It's about 5-10 times faster than the old location-based
view, but I think it can be even better. Once we have the key
dates for all the eggs from a pair, it should be pretty trivial
to figure out which birds exist and how old they are at any
date, without having to hit the database again.