basic location views added
This addresses issue #38 but some additional tuning may be needed. The location
list uses the birds() method on the Location model to query birds, which is fine
for a single location but slow when there are a lot of locations. A more
efficient approach is to use the with_locations() queryset method and then
tabulate. This is what the location_summary view does, but a downside here is
that the annotation uses a subquery that can only get one field, so we lose the
id and can't link to the individual location views. Another downside is that
we'd be paginating through a python-constructed sequence.

One option is to prefetch the event_set, limiting to the most recent event with
a location. Another is to annotate with a JSONData field, but this significantly
complicates downstream use, because we have to check if the whole field is None
before accessing the sub-fields.

It might be fine to just live with it. Each page of 25 locations takes about a
second to render which is pretty okay. Will leave the issue open for now.